:wrench: update :tada:

+++ kelm.sh
@@ -7,12 +7,46 @@ initialize () {
   MSG_LEVEL=""
   MSG_TXT=""
   DRY_RUN=""
-  FLUG_ALL_DRY_RUN=false
+  ACTION="apply"
+  FLUG_ALL=false
   FLUG_ERROR=false
+  FLUG_FILE_DELETE=true
 }
 
 show_help () {
-  echo "kload.sh -e ENV_NAME -d DIRECTORY_NAME [-p RESTART_POD_LABEL] [-c --dry-run]"
+  cat << EOS
+
+    BASICS
+      kelm -e ENV_NAME -d DIRECTORY_NAME
+    
+    OPTION
+      -e  ENV_NAME          : select values files. ex) dev-prometheus.yaml > 'dev'
+      -d  DIRECTORY_NAME    : apply directory path.
+      -p  RESTART_POD_LABEL : restart pod. require pod 'app' label value.
+      -D                    : delete manifests.
+      -c                    : enable kubectl '--dry-run' option.
+      -A                    : apply current directores all (not require -d directory).
+      -f                    : Do not delete the output manifest file.
+
+    EXAMPLE
+      # apply dev prometheus manifests.
+        > kelm -e dev -d prometheus
+
+      # apply dev prometheus manifests, and restart app=prometheus pod.
+        > kelm -e dev -d prometheus -p prometheus
+
+      # dry run stg prometheus manifests
+        > kelm -c -e stg -d prometheus
+
+      # apply dev current directory manifests.
+        > kelm -Ae dev
+
+      # dry run prd current directory manifests.
+        > kelm -Ace prd
+
+
+EOS
+
   exit 1
 }
 
@@ -48,26 +82,29 @@ if [ $# -le 1 ]; then
   show_help
 fi
 
-while getopts cChe:d:p: OPT
+while getopts De:d:p:fAch OPT
 do
   case $OPT in
     "e" )
       ENV_NAME=$OPTARG ;;
     "d" )
       DIRECTORY_NAME=$OPTARG ;;
+    "D" )
+      ACTION="delete" ;;
     "p" )
       POD_NAME=$OPTARG ;;
+    "f" )
+      FLUG_FILE_DELETE=false ;;
+    "A" )
+      FLUG_ALL=true ;;
     "c" )
       DRY_RUN="--dry-run" ;;
-    "C" )
-      DRY_RUN="--dry-run"
-      FLUG_ALL_DRY_RUN=true ;;
     "h" | * | "" )
       show_help ;;
    esac
 done
 
-if [ "$FLUG_ALL_DRY_RUN" = false ] && ( [ -z $ENV_NAME ] || [ -z $DIRECTORY_NAME ] ); then
+if [ "$FLUG_ALL" = false ] && ( [ -z $ENV_NAME ] || [ -z $DIRECTORY_NAME ] ); then
   msg "Not Found 'ENV_NAME' or 'DIRECTORY_NAME'" "e";
   exit 1
 fi
@@ -81,11 +118,19 @@ fi
 # $2 : directory
 apply_template () {
     helm template ${2} --values ${2}/values/${1}-${2}.yaml > ${1}-${2}.yaml
-    kubectl apply -f ${1}-${2}.yaml ${DRY_RUN}
-    rm -rf ${1}-${2}.yaml
+    # echo "### --- ${1}-${2} ---"
+    echo -e "\033[0;32m### --- ${1}-${2} ---\033[0;39m"
+    kubectl ${ACTION} -f ${1}-${2}.yaml ${DRY_RUN}
+    echo ""
+
+    if [ "$FLUG_FILE_DELETE" = true ]; then
+      rm -rf ${1}-${2}.yaml
+    fi
 }
 
-if [ "$FLUG_ALL_DRY_RUN" = true ]; then
+clear
+
+if [ "$FLUG_ALL" = true ]; then
   for DIR in `find . -maxdepth 1 -mindepth 1 -type d| gawk -F/ '{print $NF}'`; do
     apply_template $ENV_NAME $DIR
   done
